Write-through cache + moving getOrFetchValue
