+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/CacheInstance';
 export * from './lib/LocalCache';
 export * from './lib/RedisCache';
+export * from './lib/WriteThroughCache';
 export * from './lib/Cachette';
